Initial code

+++ src/main/java/net/argus/graphic/GraphicsListener.java
@@ -2,6 +2,6 @@ package net.argus.graphic;
 
 public interface GraphicsListener {
 	
-	public int fonction(int x);
+	public float fonction(float x);
 
 }
